+++ pkg/services/auth/authimpl/auth_token.go
@@ -1,4 +1,4 @@
-package auth
+package authimpl
 
 import (
 	"context"
